Removing Excess Ships, Implementing timed updates
Lots of ships which were time consuming to maintain correct code for
when they weren't needed for testing have been deleted; can be put in
when ready later, a few at a time. Added a io package which will
eventually have classes for graphics window for a game display. Added a
time package which controls how systems and battles will update over
time, Systems now need to all implement an update method in order to
work with this now.